requirements.txt
Django==1.11.7
django-db-file-storage==0.4.3
djangorestframework==3.7.3
gunicorn==19.7.1
psycopg2==2.7.3.2
whitenoise==3.3.1
dj_database_url==0.4.2
mysqlclient==1.3.12
hurry.filesize
django-multiupload
pyfcm
googlemaps